tests/wpt/web-platform-tests/html/semantics/forms/the-button-element/button-labels.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <p>
    <h3>button_labels</h3>
  </p>

  <hr>

  <div id="log"></div>

  <form method="post"
      enctype="application/x-www-form-urlencoded"
      action=""
      id="input_form">
    <p><label>Full name:<label>(name)<button id='button_id1'>button1</button><small>Format: First Last</small></label></label></p>
    <p><label>Age: <button id='button_id2'>button2</button></label></p>
  </form>
  <script>
    test(function() {
      var button1 = document.getElementById("button_id1");
      var button2 = document.getElementById("button_id2");

      assert_true(button1.labels instanceof NodeList, "button1.labels is NodeList");
      assert_equals(button1.labels.length, 2, "button1.labels.length");

      assert_true(button2.labels instanceof NodeList, "button2.labels is NodeList");
      assert_equals(button2.labels.length, 1, "button2.labels.length");
    });
  </script>

 </body>
</html>
